backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
boto3==1.35.86
PyJWT==2.10.1
bcrypt==4.2.1
python-multipart==0.0.20
email-validator==2.2.0
